requirements.txt
plaid-python
requests
toml
schedule